- Move Extension annotation from instance holder constant to class definition of Descriptor - Extract Descriptor as top-level class
