+++ src/main/java/hudson/plugins/plot/PlotPublisher.java
@@ -4,8 +4,6 @@
  */
 package hudson.plugins.plot;
 
-import hudson.Extension;
-import hudson.FilePath;
 import hudson.Launcher;
 import hudson.Util;
 import hudson.model.Action;
@@ -16,7 +14,6 @@ import hudson.tasks.BuildStepDescriptor;
 import hudson.tasks.BuildStepMonitor;
 import hudson.tasks.Publisher;
 import hudson.tasks.Recorder;
-import hudson.util.FormValidation;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -24,11 +21,7 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.logging.Logger;
 
-import net.sf.json.JSONObject;
 
-import org.kohsuke.stapler.AncestorInPath;
-import org.kohsuke.stapler.QueryParameter;
-import org.kohsuke.stapler.StaplerRequest;
 
 /**
  * Records the plot data for builds.
@@ -190,48 +183,5 @@ public class PlotPublisher extends Recorder {
         // misconfigured plots will not fail a build so always return true
         return true;
     }
-    @Extension
-    public static final DescriptorImpl DESCRIPTOR = new DescriptorImpl();
-
-    public static class DescriptorImpl extends BuildStepDescriptor<Publisher> {
-
-        public DescriptorImpl() {
-            super(PlotPublisher.class);
-        }
-
-        public String getDisplayName() {
-            return Messages.Plot_Publisher_DisplayName();
-        }
-
-        @Override
-        public boolean isApplicable(Class<? extends AbstractProject> jobType) {
-            return true;
-        }
-
-        /**
-         * Called when the user saves the project configuration.
-         */
-        @Override
-        public Publisher newInstance(StaplerRequest req, JSONObject formData) throws FormException {
-            PlotPublisher publisher = new PlotPublisher();
-            for (Object data : SeriesFactory.getArray(formData.get("plots"))) {
-                publisher.addPlot(bindPlot((JSONObject) data, req));
-            }
-            return publisher;
-        }
-
-        private static Plot bindPlot(JSONObject data, StaplerRequest req) {
-            Plot p = req.bindJSON(Plot.class, data);
-            p.series = SeriesFactory.createSeriesList(data.get("series"), req);
-            return p;
-        }
-
-        /**
-         * Checks if the series file is valid.
-         */
-        public FormValidation doCheckSeriesFile(@AncestorInPath AbstractProject project,
-                @QueryParameter String value) throws IOException {
-            return FilePath.validateFileMask(project.getSomeWorkspace(), value);
-        }
-    }
+    public static final PlotDescriptor DESCRIPTOR = new PlotDescriptor();
 }
